Rename legacy

+++ src/main/java/net/sodiumzh/nff/girls/entity/INFFGirlsTamed.java
@@ -152,19 +152,19 @@ public interface INFFGirlsTamed extends INFFTamed, /*IBaubleEquipable, */IAttrib
 	}
 	
 	@DontOverride
-	public default float getXpLevel()
+	public default int getXpLevel()
 	{
 		return this.getLevelHandler().getExpectedLevel();
 	}
 	
 	@DontOverride
-	public default float getOverallXp()
+	public default long getOverallXp()
 	{
 		return this.getLevelHandler().getExp();
 	}
 	
 	@DontOverride
-	public default float getXpInThisLevel()
+	public default long getXpInThisLevel()
 	{
 		return this.getLevelHandler().getExpInThisLevel();
 	}
